dynamic creation of streams and cublas_handles, support multiple streams per thread per gpu
Summary:
Currently CUDAContext only supports one cuda stream per gpu per thread. But as per my investigation, it is much better to use one CPU thread to control all streams for one GPU. To make this possible, this ground work is necessary: this diff defines a stream id for cuda context that is used to index to streams for that gpu for that thread (the streams are handled by a thread-local class).

This diff also changes the initialization: before we created cuda streams for all gpus and for all threads, even if they would be never used. Now streams are created only when needed.

This adds a small overhead to context.cuda_stream(), but I doubt that to have any significance. Instread, this diff will reduce memory usage on GPU side slightly.

Reviewed By: Yangqing

Differential Revision: D4492380

fbshipit-source-id: 498555e58d75217d43891e1bcad6d86051d376ce