Enable Reshape to handle scalars
Summary:
The context here is that we want fblearner predictor to handle float features (D4601334).
Since predictor processes a single example at a time, it makes sense to specify a single
float feature as a float scalar tensor.
But if the Caffe2 net has a SigridTransforms operator, it expects everything to have an
addition dimension so it can be called with multiple examples.

Being able to Reshape a scalar into a 1-d tensor will enable us to mix SigridTransforms
with other native Caffe2 operators.

Reviewed By: ender-wieczorek

Differential Revision: D4602675

fbshipit-source-id: 8b33876bf47bc341385fd7ac19cd1fd7f67a7ccf