increase QPS to 470K (from 250K or so)
Summary:
(Stacked with D4553941). Using the new net type increases QPS to 470K, close to Torch numbers (there are other optimizations that need to be done, particularly the log-estimator). Previously, QPS was close to 250K. This was when having reuseData=true.

Includes a small bug-fix to the new net type.

Differential Revision: D4594704

fbshipit-source-id: 21e7b0ca4173b036f45d3ba95c218792b31e7398