ability to disable inputs for extract predictor net
Summary:
This is not a super-elegant, but a working solution to fix Newsfeed-teams problem of extracting a predictor net of a net that has a "side chain" that they want to cut from the middle.

This adds a argument to ExtractPredictorModel that allows one to define "disabled inputs". These are inputs that we want to switch off, so that all operators that depend on that input will be removed from the model.

Differential Revision: D4839953

fbshipit-source-id: 5d16df6f0ec4aac6670e6917efc77abde5d75c95