special executor class for RecurrentNetworks (just single threaded now)
Summary:
This is preamble for the "diagonal executor". Instead of creating a Net for each timestep, we have a single executor for the RecurrentNetworkOp that manages ops per timestep.
This will be used if net_type='rnn', so one can still use the old way by using a net type of 'simple' or 'dag' (so there is effective kill-switch if there are some issues with this).

Did this only for the forward-model. Gradient op will follow later on, but it is basically similar, just reverse order.

Reviewed By: salexspb

Differential Revision: D4979933

fbshipit-source-id: bda77918ec518cb6b29d7021ee036d59eb2dd303