Use scope name for dataset cursor
Summary: Currently the dataset cursor blob is using a fixed name. When we read from multi input tables, the dataset cursor of each table is using the same blob. This messed up the split queue and crashed the reader pipelines (see the errors and failures in https://fb.quip.com/uzbIA7K0PgVe)

Reviewed By: dragonxlwang, rayleichen

Differential Revision: D5419863

fbshipit-source-id: 5983a3d8d2e286dc47c2ec38ed1dbbe30c7c9b49