add regulariztion in caffe2 and dper
Summary:
Regularization added for caffe2 and dper.

This regularization is intended for `dense feature `only. Sparse feature would serve as individual optimizer, see ` D5618405 ` and  `D5534579` for details.

The implementation of dense regularization is similar to the ones in optimizer. we now support `l1 norm` and  ` l2 norm` in regularizer. In dper, we would call different regularization based on regularization type defined in model_definition.thrift.

Reviewed By: xianjiec

Differential Revision: D5724851

fbshipit-source-id: 0fbee698cfeff1ac477fc9d07785406069f8d9c8