Added MAGMA_HOME env var to specify alternative MAGMA root directory (#3809)
FindMAGMA.cmake will look for MAGMA library under harcoded
/usr/local/magma by default. This commit adds MAGMA_HOME env variable
as alternative way to provide the MAGMA home directory. This is
very useful (and the only way) when the user is with restricted rights
and cannot install magma librairies under /usr/local/magma. Also it
is helpful when having multiple versions of the library and being able
to select the one to use.