Make use of new BUILD_ENVIRONMENT variable when possible. (#5699)
* Make use of new BUILD_ENVIRONMENT variable when possible.

Eliminate CI provided environment variables. At the moment, our build scripts depend on a few environment variables which are specified by the CI system and passed down to the build. Based on the build scripts, these environment variables are JOB_NAME, PYTHON_VERSION and GCC_VERSION; variables that depend solely on the image being built and the invoked script.
a. Proposal: A recent rewrite of the pytorch-dockerfiles has embedded a new environment variable, BUILD_ENVIRONMENT, which is automatically set when you run the Docker image. This environment variable subsumes JOB_NAME (this variable doesn't specify if you are “building” or “testing”, but this can easily be inferred from the script that is being invoked.) Make use of this environment variable to compute the other variables.


Signed-off-by: Internal Developer <internal-developer@intel.com>

* syntaxfix

Signed-off-by: Internal Developer <internal-developer@intel.com>

* bugfix

Signed-off-by: Internal Developer <internal-developer@intel.com>