Don't initialize the current device in CUDAGenerator::CUDAGenerator (#7392)
Previously, CUDAGenerator::CUDAGenerator would initialize the random
number generator on the current device. This would usually be device 0.
This is undesirable because initialize the CUDA context allocates a few
100 MBs due to all the kernels in libTHC.so.

This avoids the unecessary call to THCRandom_getGenerator() in the
CUDAGenerator constructor.

Fixes #7320

Previously, CUDAGenerator::CUDAGenerator would initialize the random
number generator on the current device. This would usually be device 0.
This is undesirable because initialize the CUDA context allocates a few
100 MBs due to all the kernels in libTHC.so.

This avoids the unecessary call to THCRandom_getGenerator() in the
CUDAGenerator constructor.

Fixes #7320

* Fix call to get THCState