detach spectral norm calculated weight in eval mode (#9020)
Summary:
As we left weight to be the last calculated weight in eval mode, we need to detach it from the computation in order to facilitate using backward.
The typical use case is in GANs when the discriminator has spectral norm, is in eval mode and we want to backprop through the discriminator to get weight gradients for the generator.
Closes https://github.com/pytorch/pytorch/pull/9020

Reviewed By: ezyang

Differential Revision: D8694054

Pulled By: SsnL

fbshipit-source-id: 09ee5843687cac3ed4c40759ac577a14c5371730