Add ideep copy for TensorCPU<long> in IDEEPFallbackOp (#9480)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/9480

Ops like Reshape sometimes take a second input tensor of long with the new
shape (can also be specified in arg). If this input tensor is passed in via
external input (which ONNX does sometimes), LoadOp fails with an exception.

Such ops anyway are executed by IDEEPFallbackOp, so this should be fine.

Reviewed By: yinghai

Differential Revision: D8872671

fbshipit-source-id: 659a02416c374e373ce041a7d65a174be828702d