add docs for using `python setup.py clean` in developing mode (#9524)
Summary:
This command (suggested by albanD when I raised a related question in pytorch slack) is super useful to me. I have used it several times and it worked like a charm (without it, I have to delete entire pytorch folder and clone things again). So I guess it is nice to have in the CONTRIBUTING doc.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/9524

Differential Revision: D8890126

Pulled By: soumith

fbshipit-source-id: c1798ff1ab2423627fcd8e0662a66c4e85cb2413