Use SFINAE instead of macros for 'long' hack (#12424)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/12424

Some compilers define 'long' as a separate type from 'int32_t' or 'int64_t', some don't.
Before, we had a cmake check setting a macro and depending on the macro, we either defined a separate type id for long or didn't.
Then, we removed the cmake check and used compiler detection macros directly. This is, however, error prone.

This new approach uses SFINAE to register a type id for 'long' only if it's a separate type.

Reviewed By: Yangqing, dzhulgakov

Differential Revision: D10209620

fbshipit-source-id: 68f09339e279a9a56b95caeef582c557371b518d