kill TK_NAMED_TUPLE_DEF (#24350)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/24350

`TK_NAMED_TUPLE_DEF` shouldn't exist, because NamedTuples are not
distinct syntactic things. The fact that NamedTuples and Classes are
treated differently is a property of our implementation, not the
language grammar.

This PR kills it and re-uses `CLASS_DEF` instead.

Test Plan: Imported from OSS

Differential Revision: D16825273

Pulled By: suo

fbshipit-source-id: f6d97d7e4fbdf789fd777f514eac97f32e2bbae2