Fix named tensor printing (#25564)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25564

There are a number of ops that get called while printing tensors
depending on how large the tensors are. This PR makes it so that before
we attempt to format tensor data for printing, we drop the names of the
tensor (if there are any). This is easier than supporting named tensors
for all of those ops (which should happen eventually).

Test Plan: - new test [namedtensor ci]

Differential Revision: D17158872

Pulled By: zou3519

fbshipit-source-id: 282023837645b8cb16a4d93896a843dd598fc738