Implement resize_, resize_as_ for named tensors (#26493)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26493

resize_ and resize_as_ are low level functions that are not meant to be
used as a part of the regular PyTorch user's routine. However, they are
used to implement a lot of our operations: `out=` functionality is
implemented by resizing an output to be the correct size.

To keep in line with already implemented `out=` functionality, we do the
following:
- resize_as_(self, other) propagates names according to `out=` functionality.
This means that if self doesn't have names, then we propagate
other.names. If self does have names, they must be equal to other.names.

In addition, resize_ cannot resize a named tensor to anything but the same size.

Test Plan: - [namedtensor ci]

Differential Revision: D17501404

Pulled By: zou3519

fbshipit-source-id: e396e7fba55e1419355933925226d02dccb03012