Export weight_norm (#28618)
Summary:
Export _weight_norm
Caffe2 tests are inplace

Looks like there is a conflicting behavior in torch.nn.utils.weight_norm regarding None dim.
Where dim could be negative for backwards axes, but when dim = None, it's overwitten to -1
https://github.com/pytorch/pytorch/blob/0c48092b2270d56cdab327bd1ff0ca89f5b7d569/torch/nn/utils/weight_norm.py#L10

For now, this symbolic to matches the current behavior. But this might need to be changed in the torch module.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28618

Reviewed By: hl475

Differential Revision: D18354270

Pulled By: houseroad

fbshipit-source-id: 0d64ee9ee1156bb96d36ed0a25b2e8cc5058ce90