Disable implicit conversion warning (#30529)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30529

We started to see build failures for multiple services with top-of-trunk LLVM compiler. The failures point to a warning that was treated as error for implicit conversion from long to double. Per discussion on D18642524, I'm disabling this warning from the containing TARGET file. T58053069 opened for code owner to track this - a proper source code fix and more unit test is needed.

Test Plan: local build, sandcastle

Reviewed By: smessmer

Differential Revision: D18668396

fbshipit-source-id: 28c0ff3258c5ba3afd41a0053f9fe1b356a496a8