improve the doc of enforce_sorted in pack_padded_sequence (#33617)
Summary:
this is a follow up PR to https://github.com/pytorch/pytorch/issues/33602:

torch/nn/utils/rnn.html:

`pack_padded_sequence` has a confusing and incomplete description of the `enforce_sorted` param. Currently it goes:

```
        enforce_sorted (bool, optional): if ``True``, the input is expected to
            contain sequences sorted by length in a decreasing order. If
            ``False``, this condition is not checked. Default: ``True``.
```

The second part "this condition is not checked" (1) makes no sense since the alluded to condition is not described and (2) it's incomplete as it doesn't reflect the important part, that it actually does the sorting. I think it should say something like:

```
        enforce_sorted (bool, optional): if ``True``, the input is expected to
            contain sequences sorted by length in a decreasing order. If
            ``False``, the input will get sorted unconditionally. Default: ``True``.
```
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33617

Differential Revision: D20035131

Pulled By: albanD

fbshipit-source-id: 654382eb0cb62b5abc78497faa5b4bca42db5fda