Stop generating out full function type for registration, use decltype or infer it (#33097)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33097

Previously, we had to specify full types because the functions we registering
might be overloaded, and the type was necessary to resolve the ambiguity.  I
disambiguate all of these names by mangling the names of the methods we
place on CPUType/CUDAType/TypeDefault with the overload name (these are
*internal* wrappers which are not user visible), and then can strip
the generation of full function types from the registration.

Signed-off-by: Internal Developer <internal-developer@intel.com>

Test Plan: Imported from OSS

Differential Revision: D19837898

Pulled By: ezyang

fbshipit-source-id: 5f557184f6ec84cb0613d4eb2e33b83fd1712090