.circleci: Divert packages to test channel on tag (#33842)
Summary:
This sets up PIP_UPLOAD_FOLDER to point to the correct channel for
release candidates as opposed to nightlies.

Removes an old safety check that's not needed anymore for devtoolset3

And provides a nice default for PIP_UPLOAD_FOLDER, which should clear up
confusion on where it's initially set

This is a stepping stone towards the promotable pipeline.

Signed-off-by: Internal Developer <internal-developer@intel.com>
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33842

Differential Revision: D20130791

Pulled By: seemethere

fbshipit-source-id: dac94ef46299574c36c08c968dd36faddeae6363