.circleci: Remove trailing slash, fix conda upload (#33903)
Summary:
Conda registers a suffixed slash as a new user so it was failing to
upload the anaconda packages.

In the future this should be handled through a single variable that can
be used for both but until then this will have to do.

Bug was introduced in https://github.com/pytorch/pytorch/issues/33842

Signed-off-by: Internal Developer <internal-developer@intel.com>
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33903

Differential Revision: D20148679

Pulled By: seemethere

fbshipit-source-id: 27c95f5d906ce84aa34bf5d76fd6f1ef5df08fb9