add missing 'import warnings' (#35313)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35313

The intention of D16955662 was to print a warning when a single-layer LSTM has an (ignored) dropout specified. I ran into this warning with one of our models, but instead of a warning I got "name 'warnings' is not defined". The linter could have called out that problem on the original diff, not sure why it didn't.

Test Plan: Before this diff JITing a particular model in f176977725 yielded "name 'warnings' is not defined". After this diff f176980937 gets past that point (failing in an unrelated downstream workflow).

Reviewed By: jianyuh

Differential Revision: D20611822

fbshipit-source-id: 99d90f4830f3b15ddbf1e2146e2cc014ef26c2ab