reenable quantized test_compare_tensor_scalar (#37422)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37422

The test was failing because in fbcode the version of hypothesis was too old to know
about the width parameter, and it was trying to generate values larger than float32.  The fix
is to explicitly set the defaults of the floats range for old versions of hypothesis.

For now, reenable the test and see what breaks in CI
ghstack-source-id: 103500358

Test Plan:
CI

```
buck test mode/dev-nosan //caffe2/test:quantization -- 'test_compare_tensor_scalar \(quantization\.test_quantized_op\.TestComparatorOps\)'
```

Imported from OSS

Reviewed By: raghuramank100

Differential Revision: D21310336

fbshipit-source-id: 1a59ab722daa28aab3d6d2d09bc527874942dc36