Eliminate TensorIterator::add_output with explicit dtype. (#39800)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39800

I'm working on a refactor where I want to represent the inputs
and outputs to TensorIterator as just plain Tensors, which means
I need to kill this add_output with explicit dtype.  This exists
solely to set what the output dtype should be.  We have a pretty
similar API for doing this for shapes (declare_static_shape) so
I just copied this API for dtypes instead.

Although the new version is more code, I think the intent is more
explicit.

Signed-off-by: Internal Developer <internal-developer@intel.com>

Test Plan: Imported from OSS

Differential Revision: D21981740

Pulled By: ezyang

fbshipit-source-id: cf45a6dbab6fb979ca3b412c31eca3dd4f4067de