Moving ProcessGroupNCCLTest to Gtest (#42208)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42208

ProcessGroupNCCLTest is currently written without any testing framework, and all tests are simply called from the main function and throw exceptions upon failure. As a result, it is hard to debug and pinpoint which tests have succeeded/failed.

This PR moves ProcessGroupNCCLTest to gtest with appropriate setup and skipping functionality in the test superclass.
ghstack-source-id: 109097246

Test Plan: Working Correctly on devGPU and devvm.

Reviewed By: jiayisuse

Differential Revision: D22782673

fbshipit-source-id: 85bd407f4534f3d339ddcdd65ef3d2022aeb7064