.circleci: Prefer using env-file for docker run (#43293)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43293

'docker run' has the capability to use a file for environment variables,
we should prefer to use that instead of having it be sourced per command
in the docker container.

Also opens the door for cutting down on the total number of commands we
need to echo into a script to then execute as a 'docker exec' command.

Plus side of this approach is that the BASH_ENV is persisted through all
of the steps so there's no need to do any exports / worry about
environment variables not persisting through jobs.

Signed-off-by: Internal Developer <internal-developer@intel.com>

Test Plan: Imported from OSS

Reviewed By: ezyang

Differential Revision: D23227059

Pulled By: seemethere

fbshipit-source-id: be425aa21b420b9c6e96df8b2177f508ee641a20