Switch all Sequences in tools.codegen.model to Tuple (#45127)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45127

I thought I was being clever by using Sequence, which doesn't commit to
List or Tuple, but forces read-onlyness in the type system.  However,
there is runtime implication to using List or Tuple: Lists can't be
hashed, but Tuples can be!  This is important because I shortly want
to group by FunctionSchema, and to do this I need FunctionSchema to
be hashable.  Switch everything to Tuple for true immutability.

Signed-off-by: Internal Developer <internal-developer@intel.com>

Test Plan: Imported from OSS

Reviewed By: gchanan

Differential Revision: D23872527

Pulled By: ezyang

fbshipit-source-id: 5c8fae1c50a5ae47b4167543646d94ddcafff8c3