fix: Fixed typing of bool in _ConvNd (#46828)
Summary:
Hello there :wave:

I do believe there is some typo in the typing of the `bool` argument of `_ConvNd`constructor.
The typing of the attribute is correct, but the constructor argument, while being the same way, is not the value that will be assigned to `self.bias`.

This PR simply corrects that.

Any feedback is welcome!

Pull Request resolved: https://github.com/pytorch/pytorch/pull/46828

Reviewed By: izdeby

Differential Revision: D24550435

Pulled By: ezyang

fbshipit-source-id: ab10f1a5b29a912cb23fc321a51e78b04a8391e3