Delete TypeDefault call code generation logic in VariableType (#47000)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/47000

There is a new invariant that emit_body is only ever called when
strategy is 'use_derived', which means we can delete a bunch of code.
This removes the last use of TypeXXX.h headers.

Note that this change makes sense, as the TypeDefault entries are
registered as Math entries, which means they automatically populate
Autograd (and we no longer have to register them ourselves).  Ailing
did all the hard work, this is just the payoff.

Signed-off-by: Internal Developer <internal-developer@intel.com>

Test Plan: Imported from OSS

Reviewed By: albanD

Differential Revision: D24596584

Pulled By: ezyang

fbshipit-source-id: 6fa754b5f16e75cf2dcbf437887c0fdfda5e44b1