[ROCm] add 4.1 docker image (#54628)
Summary:
Add a ROCm 4.1 docker image for CI.  Plan is to keep two ROCm versions at a time, however we still need the 3.9 image due to some CI jobs depending on it.  Keep the 4.0.1 and 3.10 images, in addition to the 3.9 image until the 3.9 image is no longer needed.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/54628

Reviewed By: H-Huang

Differential Revision: D27443378

Pulled By: malfet

fbshipit-source-id: 3f3417ec4822c6ef4c10ce2144a5b2957503dfbe