[A*][Gen-1.5] Add shape inference func for PredictorCall.
Summary:
ATT, so that the shape inference works for a model with only distributed parts.

Previously, we rely on a full_predictor net to do shape inference. For very large models, the full_predictor net won't be generated, so we have to do shape inference based on distributed parts. Surprisingly, the PredictorCall op does tensor name mapping so it has to have shape inference func supported.

Test Plan: Added unittests.

Reviewed By: khabinov

Differential Revision: D27250956

fbshipit-source-id: 3ebd36ba1eb020bb5d00358cffb8f038a6a996e8