Delete grandfathered Caffe2 dispatch keys. (#56939)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/56939

These never have kernels registered to them and are effectively useless.
What I am not so sure if we allocate tensors to them or not; if we do
I cannot use asserts and I need to ensure we just return undefined
or something equivalent.

Signed-off-by: Internal Developer <internal-developer@intel.com>

Test Plan: Imported from OSS

Reviewed By: ailzhang

Differential Revision: D28006160

Pulled By: ezyang

fbshipit-source-id: f8e2b61b8bd928fb2c0ac0b534bd4af076423f71