Modified batched residuals return of torch.linalg.lstsq (#54722)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54722

SciPy and NumPy operate only on non-batched input and return an empty array with shape (0,) if rank(a) != n.
The behavior for non-batched inputs is NumPy and SciPy compatible and the same result is computed.
For batched inputs, if any matrix in the batch has a rank less than `n`, then an empty tensor is returned.

Test Plan: Imported from OSS

Reviewed By: ngimel

Differential Revision: D27993736

Pulled By: mruberry

fbshipit-source-id: 0d7cff967b322a5e816a23f282b6ce383c4468ef