Add .isort.cfg (#60119)
Summary:
This adds the `.isort.cfg` file from https://github.com/pytorch/pytorch/issues/55928, but doesn't try to enforce it in CI because as that PR showed, that is currently difficult to do. We could use this to gradually sort the codebase according to this configuration (enforcing bits and pieces in CI) but I don't do that here.

The advantage of including this file (even if we don't enforce it) is that it affects how certain tools work, thus encouraging a specific import style for people who happen to use those tools.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/60119

Test Plan: Open `test/run_test.py` in VS Code and run the **Python Refactor: Sort Imports** command. Compare with and without this PR.

Reviewed By: 1ntEgr8

Differential Revision: D29199504

Pulled By: samestep

fbshipit-source-id: 83e937b0f517c60e3e7dedb6c0306173908fbbb0