[quant] Remove hypothesis from qtopk (#66158)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66158

qtopk used hypothesis which created flaky tests. In addition to that the tests generated were not representative, and would not catch the cases that we are interested in.

This diff removes the hypothesis from the qtopk and merges the qtopk and qtopk_nhwc tests. We now use specific testcases.
ghstack-source-id: 139768865

Test Plan: `buck test mode/dev //caffe2/test:quantization -- test_qtopk`

Reviewed By: jerryzh168

Differential Revision: D31401341

fbshipit-source-id: a8fb37a7221fc43c159f34e28aa4a91ed3506944