[TensorExpr] Add a pass for trimming JIT graphs. (#66847)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66847

Trimming means that we try to remove a small portion of the graph while
keeping it valid, and we try performing this step that N times. This is
useful for debugging when we try to find a minimal example reproducing
the issue at hand.

Differential Revision:
D31751397
D31751397

Test Plan: Imported from OSS

Reviewed By: navahgar

Pulled By: ZolotukhinM

fbshipit-source-id: 07d8ba1435af8fd2d7b8cf00db6685543fe97a85