Use nvidia cuda image without cudnn for cudnn 8 and up
Use nvidia cuda image without cudnn for cudnn 8 and up.
We want to decouple the CUDA and cudnn versions so that we can evolve these version separately.
We want to use cudnn 8.3.2 for following CUDA versions 11.3, 11.5 and 11.6.
We are using Official Nvidia Cuda ubuntu image. And installing cudnn 8.3.2 on top of it.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/74545
Approved by: https://github.com/malfet