[GH1] Add FFT related changes to merge rules
This PR would allow Quansight FFT experts (in addition to metamates) to approve sparse related changes. As the fft module is not really internally used, we can start encouraging more GitHub 1st (GH1) landing for these.

This is DIFFERENT from the superuser rule because it allows non-metamates to be approvers.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/75874
Approved by: https://github.com/osalpekar