[functorch] Generalize get_decompositions handling for OpOverloadPackets (pytorch/functorch#743)
Previously, it required only a single overload be registered to
a packet, if you asked for that packet in decompositions.  But
this is not really necessary; you could easily just return decomps
for all overloads in the packet.  This is strictly more general
and I can't imagine a situation where it's not what you want.

Signed-off-by: Internal Developer <internal-developer@intel.com>