[mergebot cli] add new revert regex (#77459)
Since we added a new CLI, we have to update the regex validator for reverting. The same validation is not there for merging.

The example command we want to validate:
`@pytorchbot merge -c="ghfirst" -m="this is breaking things internally"`
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77459
Approved by: https://github.com/malfet, https://github.com/janeyx99