[symint] Fix SizesAndStridesTest to not use negative sizes/strides
With SymInt we are using the negative space of `int64_t` in our internal
representation. `SizesAndStridesTest` breaks this because it initializes
`SizesAndStrides` with negative sizes/strides. This PR fixes that.

As an aside: feels like `SizesAndStrides` (and `SymInt`) should really
take a uint64_t, but that would be BC-breaking so I don't do it here.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/77820

Approved by: https://github.com/ezyang