[vulkan] Fix GenericList constructor (#82365)
Summary:
In some configurations, this was using the initializer_list constructor
of List, which is not valid for GenericList.  Using parens forces the
copy constructor to be used instead.  (This is a pointer copy, not a
full content copy.)

Test Plan:
CI.  Built locally.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82365
Approved by: https://github.com/SS-JIA