[composite compliance] quantile and nanquantile (#81767)
Reference: #69991

This adds a new CompositeExplicit Operator `at::assert_all_true` (also exposed in Python) to check the truthiness of a tensor and throw an error based on that.

This helps us mitigate `TORCH_CHECK(t.all().item<bool>(), "err_msg")` pattern which is not composite compliant.

Using the mentioned operator, we fix `quantile` and `nanquantile` to be Composite Compliant.

Question: Should it be documented?
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81767
Approved by: https://github.com/zou3519