Remove codesigning from github actions ios build workflow (#85597)
Codesigning isn't necessary for simulator builds, and we're not running any device tests. More importantly, our dev certificate is expiring at the end of the month, and we don't have a replacement. As a result, we need to remove our job dependencies on it. This commit removes our references to it from our github CI, but a follow up PR will be needed to remove it from CircleCI workflows.

Co-authored-by: Internal Developer <internal-developer@intel.com>

Pull Request resolved: https://github.com/pytorch/pytorch/pull/85597
Approved by: https://github.com/malfet