[functorch] Fix cross to match unbatched behavior (#86926)
Fixes #83936 #83907

In #83936, I noticed that after I wrote cross, it's silently incorrect because I misunderstood what the fix to linalg was going to be. This fixes functorch to not be silently incorrect with `linalg.cross`. Since it's a silent correctness issue that I missed, I'm hoping to cherry pick it too
Pull Request resolved: https://github.com/pytorch/pytorch/pull/86926
Approved by: https://github.com/zou3519