improving torch.linalg.lstsq documentation formatting (#89013)
Fixes #80441

The highlighting in the documentation for torch.linalg.lstsq was incorrect due to a newline that sphinx doesn't parse correctly.  Instead of writing the tensors directly, I used randn to generate the tensors.  This seems to be more consistent with how other documentation is written.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/89013
Approved by: https://github.com/lezcano