Propagate guard failures to userland (#91053)
Previously we would abort() but this is annoying when you're running
pytest or something.  Don't hard crash.

It would be nice to apply this treatment to the other uses of CHECK
macro in this file, but it was just guards that was bothering me.

Signed-off-by: Internal Developer <internal-developer@intel.com>

Pull Request resolved: https://github.com/pytorch/pytorch/pull/91053
Approved by: https://github.com/jansel