Remove fake tensor cache clearing in dynamo (#93304)
Summary: We originally cleared the cache of the converter to avoid memory leaks; now that the cache uses a weak map this is no longer necessary. Clearing of the cache caused an error in an interaction with the minifier because the minifier uses delayed compilation, so the cleanup had occurred before inductor was invoked.

Test Plan: Memory regression is being checked via dashboard and on master.

Differential Revision: D42858624

Pull Request resolved: https://github.com/pytorch/pytorch/pull/93304
Approved by: https://github.com/ezyang