Remove un-used part of cuda rng state (#98787)
The comment is quite confusing as given the use of `sizeof()`, this was never backward compatible as the state is not the same size as it used to be.

Running this through CI right now. If it turns our we serialize some rng_state Tensor, I will update the set function to be BC.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/98787
Approved by: https://github.com/ngimel