Enable a number inductor of tests on CPU (#107465)
There were many test that their `_cuda` variants were not running on
cuda. I fixed a few of these, but I'm sure there's plenty more.
It'd be great to have a way to test that we're indeed compiling
something in these tests, but I don't know how to do this off the top of
my head.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/107465
Approved by: https://github.com/ezyang