Fix reduction + () + multi-level reduction optimization (#111781)
In https://github.com/pytorch/pytorch/pull/111122, an optimization is introduced for reduction() + () + multi-level reduction. In this case, we make a multi-level reduction first-level reduction ranges the same as the previous reduction ranges so that the Inductor has better chances to fuse the first reduction and the first-level reduction of the multi-level reduction kernel together.

There is a corner case that the multi-level reduction kernel has `keepdim=True`. In this case, ranges of the multi-level reduction kernel is not empty, and the dim info needs to be used to create the inner loader of the first-level reduction kernel. To keep the logic simple, for now we simply disable optimization when `keepdim=True`.

Differential Revision: [D50544876](https://our.internmc.facebook.com/intern/diff/D50544876)

Pull Request resolved: https://github.com/pytorch/pytorch/pull/111781
Approved by: https://github.com/malfet, https://github.com/jansel