[pytorch-vulkan][1/n] Enable Packing for Vulkan Tensors (#113627)
Summary:
The new implementation of mat-mul missed a critical step that does width-packing on GPU: see T169764697.

The existing implementation of mat-mul also missed a case when the "B" matrix is already in vulkan, it fails to do packing, leading to wrong results. (I have added a disabled unittest to reflect the issue).

We will take multiple steps to enable (width / height) packing and transformation between different packing on Vulkan.

This is a first diff that enable a critical toolset: It allows development to fetch values from the underlying tensor, making it possible to implement tests for the transformation shaders.

Test Plan:
P882053410,
P882053410

Differential Revision: D51291737

Pull Request resolved: https://github.com/pytorch/pytorch/pull/113627
Approved by: https://github.com/SS-JIA