[dynamo][CI] Add numpy-2.X shard (follow up) (#140586)
Fixes #107302

This is a clone and fix for #139199.

This PR is a small step for the overall NumPy 2 support.
It adds a new CI job for testing with NumPy 2 with one test file only.
More tests to be fixed and added later in follow-up pull requests.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/140586
Approved by: https://github.com/malfet

Co-authored-by: Internal Developer <internal-developer@intel.com>